+++ tests/wpunit/Resources/Design_Tokens/Projection/Projection_ProviderTest.php
@@ -3,7 +3,7 @@
 
 namespace Tests\wpunit\Resources\Design_Tokens\Projection;
 
-use KadenceWP\KadenceBlocks\Design_Tokens\Projection\Css_Var\Hooks;
+use KadenceWP\KadenceBlocks\Design_Tokens\Projection\Css_Var\Css_Builder;
 use KadenceWP\KadenceBlocks\Design_Tokens\Projection\Css_Var\Projector;
 use KadenceWP\KadenceBlocks\Design_Tokens\Registry\Token_Registry;
 use KadenceWP\KadenceBlocks\Design_Tokens\Resolver\Token_Resolver;
@@ -12,15 +12,15 @@ use Tests\Support\Classes\TestCase;
 
 final class Projection_ProviderTest extends TestCase {
 
-	private Hooks $hooks;
+	private Projector $projector;
 	private Token_Registry $registry;
 
 	protected function setUp(): void {
 		parent::setUp();
 
-		// Hooks was registered as a singleton during module bootstrap.
-		$this->hooks    = $this->container->get( Hooks::class );
-		$this->registry = $this->container->get( Token_Registry::class );
+		// Projector was registered as a singleton during module bootstrap.
+		$this->projector = $this->container->get( Projector::class );
+		$this->registry  = $this->container->get( Token_Registry::class );
 
 		// Register the KB style handles the hooks append to.
 		if ( ! wp_style_is( 'kadence-blocks-global-variables', 'registered' ) ) {
@@ -52,14 +52,14 @@ final class Projection_ProviderTest extends TestCase {
 	// ---- Front-end enqueue ---------------------------------------------------------------------------
 
 	public function testEnqueueFrontEndAppendsInlineStyleToGlobalVariablesHandle(): void {
-		$this->hooks->enqueue_front_end();
+		$this->projector->enqueue_front_end();
 
 		$inline = wp_styles()->get_data( 'kadence-blocks-global-variables', 'after' );
 
 		$this->assertNotEmpty( $inline );
 
 		$css = implode( '', (array) $inline );
-		$this->assertStringContainsString( Projector::SCOPE, $css );
+		$this->assertStringContainsString( Css_Builder::SCOPE, $css );
 		$this->assertStringContainsString( '--kb-token--', $css );
 	}
 
@@ -70,7 +70,7 @@ final class Projection_ProviderTest extends TestCase {
 		$prev    = $pagenow;
 		$pagenow = 'post.php'; // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
 
-		$this->hooks->enqueue_editor();
+		$this->projector->enqueue_editor();
 
 		$pagenow = $prev; // phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
 
@@ -79,7 +79,7 @@ final class Projection_ProviderTest extends TestCase {
 		$this->assertNotEmpty( $inline );
 
 		$css = implode( '', (array) $inline );
-		$this->assertStringContainsString( Projector::SCOPE, $css );
+		$this->assertStringContainsString( Css_Builder::SCOPE, $css );
 		$this->assertStringContainsString( '--kb-token--', $css );
 	}
 
@@ -90,7 +90,7 @@ final class Projection_ProviderTest extends TestCase {
 			'--global-palette1' => '#3182CE',
 			'--global-palette2' => '#2B6CB0',
 		];
-		$result = $this->hooks->filter_global_colors( $input );
+		$result = $this->projector->filter_global_colors( $input );
 
 		// The shipped tokens include a palette1 slot, so that key will be rewritten; others pass through.
 		$this->assertArrayHasKey( '--global-palette1', $result );
@@ -99,7 +99,7 @@ final class Projection_ProviderTest extends TestCase {
 
 	public function testFilterFontSizesPassesThroughWhenNoFontSizeTokensRegistered(): void {
 		$input  = [ 'sm' => 'clamp(0.8rem, 0.73rem + 0.217vw, 0.9rem)' ];
-		$result = $this->hooks->filter_font_sizes( $input );
+		$result = $this->projector->filter_font_sizes( $input );
 
 		// Shipped tokens have no kadence_slot font-size entries, so input passes through unchanged.
 		$this->assertSame( $input, $result );
@@ -110,8 +110,8 @@ final class Projection_ProviderTest extends TestCase {
 	public function testNothingIsEmittedWhenRegistryIsDeactivated(): void {
 		$this->registry->deactivate();
 
-		$this->hooks->enqueue_front_end();
-		$this->hooks->enqueue_editor();
+		$this->projector->enqueue_front_end();
+		$this->projector->enqueue_editor();
 
 		$front_inline  = wp_styles()->get_data( 'kadence-blocks-global-variables', 'after' );
 		$editor_inline = wp_styles()->get_data( 'kadence-blocks-global-editor-styles', 'after' );
@@ -126,7 +126,7 @@ final class Projection_ProviderTest extends TestCase {
 		$colors = [ '--global-palette1' => '#3182CE' ];
 		$sizes  = [ 'sm' => 'clamp(0.8rem, 0.73rem + 0.217vw, 0.9rem)' ];
 
-		$this->assertSame( $colors, $this->hooks->filter_global_colors( $colors ) );
-		$this->assertSame( $sizes, $this->hooks->filter_font_sizes( $sizes ) );
+		$this->assertSame( $colors, $this->projector->filter_global_colors( $colors ) );
+		$this->assertSame( $sizes, $this->projector->filter_font_sizes( $sizes ) );
 	}
 }
